requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
Flask==3.0.3
flask-accept==0.0.6
future==1.0.0
geocoder==1.38.1
gunicorn==22.0.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pluggy==1.5.0
prettytable==3.10.2
pytest==8.3.1
pytest-randomly==3.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
ratelim==0.1.6
requests==2.32.3
six==1.16.0
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.3
